% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeFiles.R
\name{mergeFiles}
\alias{mergeFiles}
\title{Merge files}
\usage{
mergeFiles(samples, file.list)
}
\arguments{
\item{samples}{A character vector of the names of the samples that are to be 
merged in file.list.}

\item{file.list}{A list of data frames imported using the LymphoSeq function
readImmunoSeq that contain the sample names that are to be merged.  "aminoAcid", 
"nucleotide", "count" and "frequencyCount" are required columns.}
}
\value{
Returns a data frame of the merged samples.  The values of count, 
frequencyCount, and estimatedNumberGenomes are aggregated.  That is, the sum 
of count and estimatedNumberGenomes columns of the merged data frame should 
equal the sum of the columns from the unmerged samples.  Likewise, the 
frequencyCount of the merged data frame should add up to 100\%.  All other 
columns in the unmerged data frames are included in the merge data frame.
}
\description{
Merges two or more sample data frames into a single data frame and aggregates 
count, frequencyCount, and estimatedNumberGenomes.
}
\examples{
file.path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq")

file.list <- readImmunoSeq(path = file.path)

TCRB_Day949_Merged <- mergeFiles(samples = c("TRB_CD4_949", 
   "TRB_CD8_949"), file.list)

# To combine the merged data frames with file.list
file.list <- c(list("TCRB_Day949_Merged" = TCRB_Day949_Merged), file.list)
}

