% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloTree.R
\name{phyloTree}
\alias{phyloTree}
\title{Create phylogenetic tree}
\usage{
phyloTree(list, sample, type = "nucleotide", method = "ClustalOmega",
  layout = "rectangular", aa.model = "JTT", nt.model = "F81")
}
\arguments{
\item{list}{A list data frames of unproductive nucleotide sequences or productive 
nucleotide or amino acid sequenes generated by the LymphoSeq function productiveSeq 
where the parameter aggregate is set to "nucleotide".  vFamilyName, dFamilyName, and 
jFamilyName are required columns.}

\item{sample}{A character vector indicating the name of the sample in the productive
sequence list.}

\item{type}{A character vector indicating whether "aminoAcid" or "nucleotide" sequences
should be aligned.  Available options are "nucleotide" and "aminoAcid".}

\item{method}{A character vector indicating the multiple sequence alignment method to 
be used.  Refer to the Bioconductor msa package for more details.  Options incude 
"ClustalW", "ClustalOmega", and "Muscle".}

\item{layout}{A character vector indicating the tree layout.  Options include 
"rectangular", "slanted", "fan", "circular", "radial" and "unrooted".}

\item{aa.model}{A character vector indicating the pairwise distance model to be used
to compare amino acid sequences.  Refer to the phangorn package for more details.
Available options incude "WAG", "JTT", "LG", "Dayhoff", "cpREV", "mtmam", "mtArt", "MtZoa",
"mtREV24", "VT","RtREV", "HIVw", "HIVb", "FLU", "Blossum62", "Dayhoff_DCMut" and "JTT_DCMut".
The default options is "JTT".}

\item{nt.model}{A character vector indicating the pairwise distance model to be used
to compare nucleotide sequences.  Refer to the phangorn package for more details.
Available options incude "JC69" or "F81".  The default options is "F81".}
}
\value{
Returns a phylogenetic tree where each leaf represents a sequence color coded by the
V, D, and J gene usage.  The number next to each leaf refers to the sequence count.  A triangle 
shaped leaf indicates the dominant sequence.  Refer to the ggtree Bioconductor package 
documentation for details on how to manipulate the tree.
}
\description{
Create a phylogenetic tree from amino acid or nucleotide CDR3 sequences in a list of 
data frames.
}
\examples{
file.path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq")

file.list <- readImmunoSeq(path = file.path)

productive.nt <- productiveSeq(file.list = file.list, aggregate = "nucleotide")

phyloTree(list = productive.nt, sample = "TCRB_Day1320_CD8_CMV", type = "nucleotide", 
         layout = "rectangular", method = "ClustalW", nt.model = "F81")

phyloTree(list = productive.nt, sample = "TCRB_Day1320_CD8_CMV", type = "aminoAcid", 
         layout = "circular", method = "ClustalOmega", nt.model = "JTT")
}

