% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignSeq.R
\name{alignSeq}
\alias{alignSeq}
\title{Align mutliple sequences}
\usage{
alignSeq(list, sample, type = "nucleotide", method = "ClustalOmega")
}
\arguments{
\item{list}{A list of data frames consisting of antigen receptor sequences 
imported by the LymphoSeq function readImmunoSeq.}

\item{sample}{A character vector indicating the name of the sample in the productive
sequence list.}

\item{type}{A character vector indicating whether "aminoAcid" or "nucleotide" sequences
should be aligned.  If "aminoAcid" is specified, then run productiveSeqs first.}

\item{method}{A character vector indicating the multiple sequence alignment method to 
be used.  Refer to the Bioconductor msa package for more details.  Options incude 
"ClustalW", "ClustalOmega", and "Muscle".}
}
\value{
Saves a pdf file of the multiple sequence alignemnt to the working directory.
}
\description{
Perform multiple sequence alignment using one of three methods and output results as a pdf file.
}
\examples{
file.path <- system.file("extdata", "TCRB_sequencing", package = "LymphoSeq")

file.list <- readImmunoSeq(path = file.path)

productive.nt <- productiveSeq(file.list = file.list, aggregate = "nucleotide")

alignSeq(list = productive.nt, sample = "TRB_Unsorted_1320", type = "nucleotide", 
         method = "ClustalW")
}

